+++ weblate/trans/models/translation.py
@@ -164,7 +164,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
         super().__init__(*args, **kwargs)
         self.stats = TranslationStats(self)
         self.addon_commit_files = []
-        self.was_new = 0
         self.reason = ""
         self._invalidate_scheduled = False
         self.update_changes = []
@@ -220,19 +219,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
                 gettext("Could not parse file %(file)s: %(error)s")
                 % {"file": self.filename, "error": str(error)}
             )
-
-    def notify_new(self, request):
-        if self.was_new:
-            # Create change after flags has been updated and cache
-            # invalidated, otherwise we might be sending notification
-            # with outdated values
-            self.change_set.create(
-                action=Change.ACTION_NEW_STRING,
-                user=request.user if request else None,
-                author=request.user if request else None,
-                details={"count": self.was_new},
-            )
-            self.was_new = 0
 
     def get_url_path(self):
         return (*self.component.get_url_path(), self.language.code)
@@ -328,18 +314,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
         ):
             newunit.update_from_unit(unit, pos, is_new)
 
-        # Check if unit is worth notification:
-        # - new and untranslated
-        # - newly untranslated
-        # - newly fuzzy
-        # - source string changed
-        if newunit.state < STATE_TRANSLATED and (
-            newunit.state != newunit.old_unit["state"]
-            or is_new
-            or newunit.source != newunit.old_unit["source"]
-        ):
-            self.was_new += 1
-
         # Store current unit ID
         updated[id_hash] = newunit
 
@@ -413,9 +387,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
                 if plural != self.plural:
                     self.plural = plural
                     self.save(update_fields=["plural"])
-
-                # Was there change?
-                self.was_new = 0
 
                 # Select all current units for update
                 dbunits = {
@@ -1171,8 +1142,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
             )
             existing.add(idkey)
             accepted += 1
-        self.was_new = accepted
-        self.notify_new(request)
         component.invalidate_cache()
         if component.needs_variants_update:
             component.update_variants()
@@ -1351,7 +1320,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
             self.component.invalidate_cache()
         else:
             self.check_sync(request=request, change=change)
-            self.notify_new(request)
             self.invalidate_cache()
         # Trigger post-update signal
         self.component.trigger_post_update(previous_revision, False)
@@ -1521,8 +1489,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
                 )
             component.invalidate_cache()
             component_post_update.send(sender=self.__class__, component=component)
-            self.was_new = 1
-            self.notify_new(request)
         return result
 
     def notify_deletion(self, unit, user):
@@ -1599,7 +1565,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
         if not self.is_source or not self.component.manage_units:
             return
         expected_count = self.component.translation_set.count()
-        self.was_new = 0
         for source in self.component.get_all_sources():
             # Is the string a terminology
             if "terminology" not in source.all_flags:
@@ -1615,8 +1580,6 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
                 is_batch_update=True,
                 skip_existing=True,
             )
-            self.was_new += 1
-        self.notify_new(None)
 
     def validate_new_unit_data(
         self,